deps: updates TinyGo to v0.28.1

+++ handler/header.go
@@ -31,8 +31,8 @@ func (w wasmHeader) Names() (names []string) {
 	}
 	// Otherwise, we have to allocate a new buffer for the large entry.
 	buf := make([]byte, size)
-	ptr := unsafe.Pointer(unsafe.SliceData(buf))
-	_ = imports.GetHeaderNames(imports.HeaderKind(w), uintptr(ptr), size)
+	ptr := uint32(uintptr(unsafe.Pointer(unsafe.SliceData(buf))))
+	_ = imports.GetHeaderNames(imports.HeaderKind(w), ptr, size)
 	names = mem.GetNULTerminated(buf)
 	runtime.KeepAlive(buf) // keep buf alive until ptr is no longer needed.
 	return
@@ -51,7 +51,7 @@ func (w wasmHeader) Get(name string) (value string, ok bool) {
 // GetAll implements the same method as documented on api.Request.
 func (w wasmHeader) GetAll(name string) (names []string) {
 	namePtr, nameSize := mem.StringToPtr(name)
-	countLen := imports.GetHeaderValues(imports.HeaderKind(w), uintptr(namePtr), nameSize, mem.ReadBufPtr, mem.ReadBufLimit)
+	countLen := imports.GetHeaderValues(imports.HeaderKind(w), namePtr, nameSize, mem.ReadBufPtr, mem.ReadBufLimit)
 	runtime.KeepAlive(name) // keep name alive until ptr is no longer needed.
 	if countLen == 0 {
 		return
@@ -66,8 +66,8 @@ func (w wasmHeader) GetAll(name string) (names []string) {
 	}
 	// Otherwise, we have to allocate a new buffer for the large entry.
 	buf := make([]byte, size)
-	ptr := unsafe.Pointer(unsafe.SliceData(buf))
-	_ = imports.GetHeaderValues(imports.HeaderKind(w), uintptr(namePtr), nameSize, uintptr(ptr), size)
+	ptr := uint32(uintptr(unsafe.Pointer(unsafe.SliceData(buf))))
+	_ = imports.GetHeaderValues(imports.HeaderKind(w), namePtr, nameSize, ptr, size)
 	names = mem.GetNULTerminated(buf)
 	runtime.KeepAlive(name) // keep name alive until ptr is no longer needed.
 	runtime.KeepAlive(buf)  // keep buf alive until ptr is no longer needed.
@@ -78,7 +78,7 @@ func (w wasmHeader) GetAll(name string) (names []string) {
 func (w wasmHeader) Set(name, value string) {
 	namePtr, nameSize := mem.StringToPtr(name)
 	valuePtr, valueSize := mem.StringToPtr(value)
-	imports.SetHeaderValue(imports.HeaderKind(w), uintptr(namePtr), nameSize, uintptr(valuePtr), valueSize)
+	imports.SetHeaderValue(imports.HeaderKind(w), namePtr, nameSize, valuePtr, valueSize)
 	runtime.KeepAlive(name)  // keep name alive until ptr is no longer needed.
 	runtime.KeepAlive(value) // keep value alive until ptr is no longer needed.
 }
@@ -87,7 +87,7 @@ func (w wasmHeader) Set(name, value string) {
 func (w wasmHeader) Add(name, value string) {
 	namePtr, nameSize := mem.StringToPtr(name)
 	valuePtr, valueSize := mem.StringToPtr(value)
-	imports.AddHeaderValue(imports.HeaderKind(w), uintptr(namePtr), nameSize, uintptr(valuePtr), valueSize)
+	imports.AddHeaderValue(imports.HeaderKind(w), namePtr, nameSize, valuePtr, valueSize)
 	runtime.KeepAlive(name)  // keep name alive until ptr is no longer needed.
 	runtime.KeepAlive(value) // keep value alive until ptr is no longer needed.
 }
@@ -95,6 +95,6 @@ func (w wasmHeader) Add(name, value string) {
 // Remove implements the same method as documented on api.Request.
 func (w wasmHeader) Remove(name string) {
 	namePtr, nameSize := mem.StringToPtr(name)
-	imports.RemoveHeader(imports.HeaderKind(w), uintptr(namePtr), nameSize)
+	imports.RemoveHeader(imports.HeaderKind(w), namePtr, nameSize)
 	runtime.KeepAlive(name) // keep name alive until ptr is no longer needed.
 }
